Config for connector
When creating test data (e.g. in test-complex_missing_data.R), this is done in a temporary directory. This is then linked to the testthat/fixtures folder with a symbolic link "data". This way, connector can be "static" and point to this link, while data is temporarily created for each test run